requirements/base.txt
asgiref==3.6.0
Django==4.2
sqlparse==0.4.4
django-cors-headers==3.14.0
django-environ==0.10.0
djangorestframework==3.14.0
Pillow==9.5.0
psycopg2-binary==2.9.6
djangorestframework-simplejwt==5.2.2